[enhancement/SOLI-principle]: enhance the use of SOLID principles and improve OOP design
