+++ assets/js/bookManager.js
@@ -5,7 +5,7 @@ export class BookManager {
     showTable;
     bookManagerHelper;
     toggleEvent;
-    constructor(utils, formhandler, bookStore, showTable, bookManagerHelper) {
+    constructor(formhandler, bookStore, showTable, bookManagerHelper) {
         this.formhandler = formhandler;
         this.bookStore = bookStore;
         this.showTable = showTable;
@@ -24,18 +24,6 @@ export class BookManager {
         genreFilter?.addEventListener('change', this.filterGenre.bind(this));
         this.bookManagerHelper.toggleSortBtn();
     }
-    // public bindTableEvents(): void {
-    //   console.log('Buttons: ', document.querySelectorAll('[id^="editBtn-"]'));
-    //   document.querySelectorAll('button[id^="editBtn-"]').forEach((btn) => {
-    //     const index = parseInt((btn as HTMLElement).id.split('-')[1]);
-    //     btn.addEventListener('click', () => this.editBook(index));
-    //   });
-    //   document.querySelectorAll('button[id^="deleteBtn-"]').forEach((btn) => {
-    //     const index = parseInt((btn as HTMLElement).id.split('-')[1]);
-    //     btn.addEventListener('click', () => this.deleteBook(index));
-    //   });
-    //   this.toggleSortBtn();
-    // }
     updateSubmitButton(text) {
         document.querySelector('[type="submit"]').textContent = text;
     }
@@ -62,7 +50,6 @@ export class BookManager {
         Utils.showModal('Book Deleted Successfully');
     }
     deleteBook(index) {
-        console.log(`Deleting book at index: ${index}`);
         this.showDeleteConfirmationModal(index);
     }
     editBook(index) {
@@ -81,38 +68,6 @@ export class BookManager {
             : this.bookStore.bookList;
         this.showTable.updateTableData(filteredBooks);
         // this.bindTableEvents();
-    }
-    discountCalculation(price, discountedPrice) {
-        if (price === undefined && discountedPrice === undefined) {
-            price = 0;
-            discountedPrice = 0;
-            if (price === discountedPrice) {
-                return `<span class="text-green-500 font-bold">${price.toFixed()} /-</span>`;
-            }
-            else {
-                const percentage = (discountedPrice / price) * 100;
-                const discountPercentage = (100 - percentage).toFixed();
-                return `
-          <span class="line-through text-red-500 font-semibold">${price.toFixed()} rs/-</span>
-          <span class="text-green-600 font-bold">(${discountPercentage}% Off)</span><br>
-          <span class=" text-blue-600  font-semibold">${discountedPrice.toFixed()} rs/-</span>
-        `;
-            }
-        }
-        else {
-            if (price === discountedPrice || discountedPrice === 0) {
-                return `<span class="text-green-500 font-bold">${price.toFixed()} rs/-</span>`;
-            }
-            else {
-                const percentage = (discountedPrice / price) * 100;
-                const discountPercentage = (100 - percentage).toFixed();
-                return `
-          <span class="line-through text-red-500 font-semibold">${price.toFixed()} rs/-</span>
-          <span class="text-green-600 font-bold">(${discountPercentage}% Off)</span><br>
-          <span class=" text-blue-600  font-semibold">${discountedPrice.toFixed()} rs/-</span>
-        `;
-            }
-        }
     }
     // Method to set form data
     setForm(data) {
