+++ assets/js/config.js
@@ -1,4 +1,3 @@
-// config.js
 export const config = {
     apiUrl: "https://www.googleapis.com/books/v1/volumes?q=",
 };
