+++ assets/ts/formHandler.ts
@@ -1,31 +1,18 @@
-import { Utils } from './utils';
-import { Book } from './script';
-
-export interface formData {
-  title: string;
-  author: string;
-  isbn: string;
-  publicationDate: string;
-  price: number;
-  discountPrice: number;
-  genre: string;
-  bookAge: string;
-}
+import { Utils } from './utils.js';
+import { formData } from './bookStore.js';
+import { BookStore } from './bookStore.js';
+import { BookManagerHelper } from './bookManageHelper.js';
 
 export class FormHandler {
   public form: HTMLFormElement | null;
-  private utils: Utils;
-  private book: Book | null;
   public cachedFormData: formData;
   public editIndex: number | null;
-  public bookList: formData[];
 
-  constructor(utils: Utils) {
+  constructor(private bookStore: BookStore, private bookManagerHelper: BookManagerHelper) {
     this.form = document.getElementById('formData') as HTMLFormElement | null;
-    this.utils = utils;
-    this.book = null;
+    this.bookManagerHelper = bookManagerHelper;
+    this.bookStore = bookStore;
     this.editIndex = null;
-    this.bookList = [];
     this.cachedFormData = {
       title: '',
       author: '',
@@ -33,19 +20,17 @@ export class FormHandler {
       publicationDate: '',
       price: 0,
       discountPrice: 0,
+      discountedPrice: '',
       genre: '',
       bookAge: '',
     };
     this.bindFormHandler();
   }
 
-  setBookInstance(book: Book): void {
-    this.book = book;
-  }
-
   bindFormHandler() {
     this.form?.addEventListener('submit', this.handleFormSubmit.bind(this));
   }
+
   // Method to handle form reset
   public handleFormReset(): void {
     this.editIndex = null;
@@ -64,12 +49,14 @@ export class FormHandler {
     ];
     errorFields.forEach((field) => Utils.toggleError(field));
   }
+
   // Update submit button text for editing or adding
   public updateSubmitButton(text: string): void {
     (
       document.querySelector('[type="submit"]') as HTMLButtonElement
     ).textContent = text;
   }
+
   // Method to handle form submission
   public handleFormSubmit(event: Event): void {
     event.preventDefault();
@@ -85,17 +72,18 @@ export class FormHandler {
       discountPrice,
       genre,
     } = this.cachedFormData;
-    const bookAge = this.calculateBookAge(publicationDate).trim();
+    const bookAge = this.bookManagerHelper.calculateBookAge(publicationDate).trim();
     if (bookAge) {
+      // if editIndex is not null, then edit the book
       if (this.editIndex !== null) {
-        // Editing a book
-        this.bookList[this.editIndex] = {
-          ...this.bookList[this.editIndex],
+        this.bookStore.bookList[this.editIndex] = {
+          ...this.bookStore.bookList[this.editIndex],
           title,
           author,
           publicationDate,
           price,
           discountPrice,
+          discountedPrice: this.bookManagerHelper.discountCalculation(price, discountPrice),
           genre,
           bookAge,
         };
@@ -103,52 +91,27 @@ export class FormHandler {
         Utils.showModal('Book Edited Successfully!');
       } else {
         // Adding a new book
-        this.bookList.push({
+        this.bookStore.bookList.unshift({
           title,
           author,
           isbn,
           publicationDate,
           price,
           discountPrice,
+          discountedPrice: this.bookManagerHelper.discountCalculation(price, discountPrice),
           genre,
           bookAge,
         });
+        this.bookManagerHelper.toggleSortBtn();
         Utils.showModal('Book Added Successfully');
       }
 
-      // Update the table data
-      this.book?.updateTableData(this.bookList);
-      // Rebind events to the table rows
-      // this.book?.bindEvents();
+      this.bookStore.displayBooksData();
       this.handleFormReset();
-      // Update total book count
-      this.book?.totalBookCount();
+      this.bookStore.totalBookCount();
     }
   }
 
-  // Method to calculate the age of the book
-  public calculateBookAge(publicationDate: string): string {
-    const currentDate = new Date();
-    const pubDate = new Date(publicationDate);
-    const ageInYears = currentDate.getFullYear() - pubDate.getFullYear();
-    const ageInMonths = currentDate.getMonth() - pubDate.getMonth();
-    const ageInDays = currentDate.getDate() - pubDate.getDate();
-
-    if (currentDate < pubDate) {
-      Utils.toggleError("publicationDate-error", "Future dates are not permissible. Please choose a valid date")
-      return '';
-    }
-
-    let ageText =
-      ageInYears > 0
-        ? `${ageInYears} year(s)`
-        : ageInMonths > 0
-        ? `${ageInMonths} month(s)`
-        : ageInDays > 0
-        ? `${ageInDays} day(s)`
-        : `Less than a day old`;
-    return ageText;
-  }
   // Get form data and return as an object
   public getFormData(): formData {
     return {
@@ -241,42 +204,5 @@ export class FormHandler {
     }
 
     return isValid;
-  }
-  // Method to set form data
-  public setForm(data: formData): void {
-    const formElements = {
-      title: document.getElementById('title') as HTMLInputElement,
-      author: document.getElementById('author') as HTMLInputElement,
-      isbn: document.getElementById('isbn') as HTMLInputElement,
-      publicationDate: document.getElementById(
-        'publicationDate',
-      ) as HTMLInputElement,
-      listPrice: document.getElementById('listPrice') as HTMLInputElement,
-      discountPrice: document.getElementById(
-        'discountPrice',
-      ) as HTMLInputElement,
-      genre: document.getElementById('genre') as HTMLSelectElement,
-    };
-    formElements.title.value = data.title;
-    formElements.author.value = data.author;
-    formElements.isbn.value = data.isbn;
-    formElements.isbn.disabled = true;
-    formElements.publicationDate.value = data.publicationDate;
-    formElements.listPrice.value = data.price.toString();
-    formElements.discountPrice.value = data.discountPrice.toString();
-
-    const genreSelect = formElements.genre;
-    const genreOptions = Array.from(genreSelect.options).map((option) =>
-      option.value.toLowerCase(),
-    );
-    const bookGenre = data.genre.trim().toLowerCase();
-    if (!genreOptions.includes(bookGenre)) {
-      const newOption = document.createElement('option');
-      newOption.value = data.genre;
-      newOption.textContent = data.genre;
-      genreSelect.appendChild(newOption);
-    }
-    genreSelect.value = data.genre;
-    formElements.title.focus();
   }
 }
