+++ assets/ts/main.ts
@@ -1,34 +1,44 @@
-import { Book } from './script';
-import { Utils } from './utils';
-import { FormHandler } from './formHandler';
-import { FetchBook } from './fetchBook';
+import { BookStore } from './bookStore.ts';
+import { BookManager } from './bookManager.ts';
+import { FormHandler } from './formHandler.ts';
+import { FetchBook } from './fetchBook.ts';
+import { ShowTable } from './showTable.ts';
+import { BookManagerHelper } from './bookManageHelper.ts';
 
-const utils = new Utils();
+document.addEventListener('DOMContentLoaded', () => {
+  const showTable = new ShowTable();
+  const bookStore = new BookStore(showTable);
+  const bookManagerHelper = new BookManagerHelper(bookStore, showTable);
+  const formHandler = new FormHandler(bookStore, bookManagerHelper);
+  const book = new BookManager(
+    formHandler,
+    bookStore,
+    showTable,
+    bookManagerHelper,
+  );
+  const fetchBook = new FetchBook(bookStore, bookManagerHelper, book);
 
-const formHandler = new FormHandler(utils);
+  const tableElement = document.getElementById('tableData') as HTMLTableElement;
+  tableElement.addEventListener('click', (e) => {
+    const target = e.target as HTMLElement;
+    if (target.matches('button[id^="editBtn-"]')) {
+      const index = parseInt(target.id.split('-')[1]);
+      book.editBook(index);
+    } else if (target.matches('button[id^="deleteBtn-"]')) {
+      const index = parseInt(target.id.split('-')[1]);
+      book.deleteBook(index);
+    }
+  });
 
-const book = new Book(utils, formHandler);
-book.bindEvents();
-formHandler.setBookInstance(book);
-book.setFormHandlerInstance(formHandler);
+  book.bindEvents();
 
-const fetchBook = new FetchBook(formHandler, book, utils);
+  (document.getElementById('searchBtn') as HTMLButtonElement).addEventListener(
+    'click',
+    (e) => fetchBook.searchBook(e),
+  );
+  (
+    document.getElementById('clearSearchBtn') as HTMLButtonElement
+  ).addEventListener('click', () => fetchBook.clearSearch());
 
-
-(document.getElementById('searchBtn') as HTMLButtonElement).addEventListener(
-  'click',
-  (e) => fetchBook.searchBook(e),
-);
-(document.getElementById('clearSearchBtn') as HTMLButtonElement).addEventListener(
-  'click',
-  () => fetchBook.clearSearch(),
-);
-
-// Pass form events to FormHandler
-(document.getElementById('form') as HTMLFormElement)?.addEventListener(
-  'submit',
-  (e) => formHandler.handleFormSubmit(e),
-);
-
-// Example usage of FetchBook
-fetchBook.fetchBooks();
+  fetchBook.fetchBooks();
+});
