+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "eslint": "^9.18.0"
+  }
+}
